Fix failing subsequent delta updates on Windows
Before this commit, Brave's full and delta installers had the same output file:
brave_installer.exe. This meant that, to create both kinds of installers, we
had to run the associated target (create_dist) twice. This had the unintended
side effect of signing the DLLs and EXEs that end up in Brave's installation
directory twice. That in turn broke subsequent delta updates because the two
singing runs produce binaries with different hashes.

For example, say chrome.dll has hash "1" after the call to create_dist that
generates the full installer, and hash "2" after the call that creates the delta
installer. When the delta installer runs, it installs chrome.dll with hash "2".
But now consider the next delta. It gets generated from the previous full
installer, where chrome.dll had hash "1". But when that second delta gets
applied, it sees the (to it) unexpected hash "2" from the first run. It fails
and aborts because of this mismatch.

This commit introduces a separate file for the delta installer:
delta_installer.exe. This allows producing the full and delta installer in one
create_dist call. The new implementation also uses the previously signed
binaries in chrome.7z instead of signing them again. At the moment, performing
a single create_dist build (with full and delta installer) fixes the problem
described above. Performing two such builds (one for full, one for delta) still
exhibits the problem because create_dist always rebuilds chrome.7z.